update estimation methods to improve accuracy for sparse data

+++ core/src/main/java/io/questdb/griffin/engine/groupby/MonthTimestampMicrosSampler.java
@@ -57,9 +57,9 @@ public class MonthTimestampMicrosSampler implements TimestampSampler {
     }
 
     @Override
-    public long nextTimestamp(long timestamp, int numSteps) {
+    public long nextTimestamp(long timestamp, long numSteps) {
         try {
-            return addMonth(timestamp, Math.multiplyExact(numSteps, stepMonths));
+            return addMonth(timestamp, (int) Math.multiplyExact(numSteps, stepMonths));
         } catch (ArithmeticException e) {
             return Long.MAX_VALUE;
         }
