+++ core/src/main/java/io/questdb/griffin/engine/groupby/YearTimestampMicrosSampler.java
@@ -56,9 +56,9 @@ public class YearTimestampMicrosSampler implements TimestampSampler {
     }
 
     @Override
-    public long nextTimestamp(long timestamp, int numSteps) {
+    public long nextTimestamp(long timestamp, long numSteps) {
         try {
-            return addYears(timestamp, Math.multiplyExact(numSteps, stepYears));
+            return addYears(timestamp, (int) Math.multiplyExact(numSteps, stepYears));
         } catch (ArithmeticException e) {
             return Long.MAX_VALUE;
         }
